Format ligand list items
 - added LigandResultItem.js
 -renamed FilteredResultsBox.js to FilteredLigandResults.js
